+++ src/tol/open_citations/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: 2026 Genome Research Ltd.
+#
+# SPDX-License-Identifier: MIT
+
+from .open_citations_datasource import *  # noqa
+from .factory import create_open_citations_datasource  # noqa